# backend/docs/swagger.yaml
definitions:
...
  models.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  models.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
...
  models.MessageResponse:
    properties:
      message:
        type: string
    type: object
...
  models.Role:
    properties:
      group:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  models.SSOAuthResponse:
    properties:
      token:
        type: string
      user:
        allOf:
        - $ref: '#/definitions/models.UserInfo'
        description: Full UserInfo with all SSO fields except password
    type: object
...
  models.UserInfo:
    properties:
      address_en:
        description: En เลขที่ อาคาร ชั้น ห้อง ชื่อหมู่บ้าน
        type: string
      address_no:
        description: เลขที่ อาคาร ชั้น ห้อง ชื่อหมู่บ้าน
        type: string
      applicanttype_id:
        description: ประเภทผู้สมัคร
        type: string
      block:
        description: 0=ไม่ถูกบล็อก, 1=ถูกบล็อก
        type: integer
      branch_code:
        description: รหัสสาขา
        type: string
      branch_type:
        description: ประเภทสาขา 1.สำนักงานใหญ่, 2.สาขา
        type: string
      building:
        description: อาคาร
        type: string
      check_api:
        description: เช็คสถานะ API 1.API
        type: integer
      contact_address_en:
        description: En เลขที่ (ที่อยู่ที่ติดต่อได้)
        type: string
      contact_address_no:
        description: เลขที่ (ที่อยู่ที่ติดต่อได้)
        type: string
      contact_building:
        description: อาคาร (ที่อยู่ที่ติดต่อได้)
        type: string
      contact_district:
        description: เขต/อำเภอ (ที่อยู่ที่ติดต่อได้)
        type: string
      contact_district_en:
        description: En เขต/อำเภอ (ที่อยู่ที่ติดต่อได้)
        type: string
      contact_fax:
        description: แฟกซ์ (ผู้ติดต่อ)
        type: string
      contact_first_name:
        description: ชื่อ (ผู้ติดต่อ)
        type: string
      contact_last_name:
        description: นามสกุล (ผู้ติดต่อ)
        type: string
      contact_moo:
        description: หมู่ (ที่อยู่ที่ติดต่อได้)
        type: string
      contact_moo_en:
        description: En หมู่ (ที่อยู่ที่ติดต่อได้)
        type: string
      contact_name:
        description: ชื่อผู้ติดต่อเจ้าของ E-mail
        type: string
      contact_phone_number:
        description: เบอร์โทรศัพท์ (ผู้ติดต่อ)
        type: string
      contact_position:
        description: ตำแหน่ง (ผู้ติดต่อ)
        type: string
      contact_prefix_name:
        description: คำนำหน้าชื่อ (ผู้ติดต่อ)
        type: string
      contact_prefix_text:
        description: คำนำหน้าชื่อ (ผู้ติดต่อ) text
        type: string
      contact_province:
        description: จังหวัด (ที่อยู่ที่ติดต่อได้)
        type: string
      contact_province_en:
        description: En จังหวัด (ที่อยู่ที่ติดต่อได้)
        type: string
      contact_soi:
        description: ตรอก/ซอย (ที่อยู่ที่ติดต่อได้)
        type: string
      contact_soi_en:
        description: En ตรอก/ซอย (ที่อยู่ที่ติดต่อได้)
        type: string
      contact_street:
        description: ถนน (ที่อยู่ที่ติดต่อได้)
        type: string
      contact_street_en:
        description: En ถนน (ที่อยู่ที่ติดต่อได้)
        type: string
      contact_subdistrict:
        description: แขวง/ตำบล (ที่อยู่ที่ติดต่อได้)
        type: string
      contact_subdistrict_en:
        description: En แขวง/ตำบล (ที่อยู่ที่ติดต่อได้)
        type: string
      contact_tax_id:
        description: เลขประจำตัวผู้เสียภาษี (ผู้ติดต่อ)
        type: string
      contact_tel:
        description: โทรศัพท์ (ผู้ติดต่อ)
        type: string
      contact_zipcode:
        description: รหัสไปรษณีย์ (ที่อยู่ที่ติดต่อได้)
        type: string
      contact_zipcode_en:
        description: En รหัสไปรษณีย์ (ที่อยู่ที่ติดต่อได้)
        type: string
      corporatefile:
        description: หนังสือรับรองหรือสำเนาใบสำคัญ
        type: string
      date_niti:
        description: วันที่นิติบุคคล
        type: string
      date_of_birth:
        description: วันเกิด
        type: string
      district:
        description: เขต/อำเภอ
        type: string
      district_en:
        description: En เขต/อำเภอ
        type: string
      email:
        type: string
      fax:
        description: แฟกซ์
        type: string
      google2fa_secret:
        description: Login 2 ขั้นตอน Google Authenticator รหัสลับ
        type: string
      google2fa_status:
        description: Login 2 ขั้นตอน Google Authenticator 0=ปิดใช้, 1=เปิดใช้
        type: integer
      id:
        description: SSO Users fields (sso_users table) - All fields except password
        type: integer
      juristic_cause_quit:
        description: สาเหตุที่เลิกกิจการ กรณี juristic_status=4
        type: string
      juristic_status:
        description: สถานะนิติบุคคล 1.ยังดำเนินกิจการอยู่, 2.ฟื้นฟู, 3.คืนสู่ทะเบียน,
          4.เลิกกิจการ
        type: integer
      lastResetTime:
        description: Date of last password reset
        type: string
      lastvisitDate:
        description: วันที่เข้าใช้งานล่าสุด
        type: string
      latitude:
        description: พิกัดที่ตั้ง (ลองจิจูด)
        type: string
      longitude:
        description: พิกัดที่ตั้ง (ละติจูด)
        type: string
      moo:
        description: หมู่
        type: string
      moo_en:
        description: En หมู่
        type: string
      name:
        description: ชื่อ/บริษัท
        type: string
      name_en:
        description: En ชื่อ/บริษัท
        type: string
      nationality:
        description: สัญชาติ
        type: string
      params:
        description: Parameters
        type: string
      password:
        type: string
      passwordHash:
        type: string
      person_first_name:
        description: ชื่อ (บุคคล)
        type: string
      person_last_name:
        description: นามสกุล (บุคคล)
        type: string
      person_type:
        description: ประเภทบุคคล
        type: string
      personfile:
        description: สำเนาบัตรประจำตัวประชาชน
        type: string
      picture:
        description: ชื่อไฟล์ภาพประจำตัว
        type: string
      prefix_name:
        description: คำนำหน้าชื่อ
        type: string
      prefix_text:
        description: คำนำหน้าชื่อ text
        type: string
      province:
        description: จังหวัด
        type: string
      province_en:
        description: En จังหวัด
        type: string
      reg_subdepart:
        type: string
      registerDate:
        description: วันที่สมัคร
        type: string
      remember_token:
        description: Remember token
        type: string
      resetCount:
        description: Count of password resets since lastResetTime
        type: integer
      role:
        $ref: '#/definitions/models.Role'
      sendEmail:
        description: ส่งอีเมล
        type: integer
      soi:
        description: ตรอก/ซอย
        type: string
      soi_en:
        description: En ตรอก/ซอย
        type: string
      state:
        description: 1=รอยืนยัน, 2=ยืนยันแล้ว, 3=รอเจ้าหน้าที่เปิดใช้งาน
        type: integer
      street:
        description: ถนน
        type: string
      street_en:
        description: En ถนน
        type: string
      subdistrict:
        description: แขวง/ตำบล
        type: string
      subdistrict_en:
        description: En แขวง/ตำบล
        type: string
      tax_number:
        description: เลขประจำตัวผู้เสียภาษี
        type: string
      tel:
        description: โทรศัพท์
        type: string
      username:
        type: string
      zipcode:
        description: รหัสไปรษณีย์
        type: string
      zipcode_en:
        description: En รหัสไปรษณีย์
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/estandard/auth/logout:
    post:
      consumes:
      - application/json
      description: Logout current user by removing token cookie
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Logout
      tags:
      - Auth
...
  /api/v1/estandard/sso/login:
    post:
      consumes:
      - application/json
      description: Profile with username or email  from sso_users table. Returns full
        user information from sso_users table (except password)
      parameters:
      - description: Login credentials (username or email)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SSOAuthResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: SSO Profile (sso_users table)
      tags:
      - Auth
  /api/v1/estandard/sso/profile-admin:
    post:
      consumes:
      - application/json
      description: Profile with username or email from user_register table. Returns
        full user information mapped to UserInfo model (except password) and token
      parameters:
      - description: Login credentials (username or email)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SSOAuthResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: SSO Profile Admin (user_register table)
      tags:
      - Auth
...
swagger: "2.0"